requirements.gpu-cu80.txt
pyyaml
mxnet-cu80mkl==0.11
numpy>=1.12
typing
